+++ CHANGELOG.md
@@ -5,4 +5,6 @@
 01. Added dev.orne.config.HierarchicalConfig
 01. Added dev.orne.config.AbstractStringConfig
 01. Added dev.orne.config.SystemConfig
+01. Added dev.orne.config.MutableConfig
+01. Added dev.orne.config.AbstractMutableStringConfig
 01. Added dev.orne.config.PropertiesConfig